// lib/api/generated/model/listAuditLogsTenantsCurrentAuditLogsGetParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Syncnesto API
 * OpenAPI spec version: 0.1.0
 */

export type ListAuditLogsTenantsCurrentAuditLogsGetParams = {
  event_type?: string | null;
  actor_user_id?: number | null;
  project_id?: number | null;
  created_from?: string | null;
  created_before?: string | null;
  /**
   * @minimum 1
   * @maximum 500
   */
  page?: number;
  /**
   * @minimum 1
   * @maximum 50
   */
  page_size?: number;
};